backend/requirements.txt
fastapi
uvicorn
pydantic
transformers
peft
torch
accelerate
python-multipart
python-dotenv